GenericRepository практически готов

+++ AurigaPetProject2023.DataAccess/Entities/ProductType.cs
@@ -5,7 +5,7 @@ namespace AurigaPetProject2023.DataAccess.Entities
     public class ProductType
     {
         [Column("ProductType_ID")]
-        public long ProductTypeID { get; set; }
+        public int ProductTypeID { get; set; }
         public string Name { get; set; }
         [Column("Unique")]
         public bool IsUnique { get; set; }
